+++ src/screen.ts
@@ -1,4 +1,3 @@
-import {stderr, stdout} from './cli-ux/stream'
 import {settings} from './settings'
 
 function termwidth(stream: any): number {
@@ -20,5 +19,5 @@ function termwidth(stream: any): number {
 
 const columns = Number.parseInt(process.env.OCLIF_COLUMNS!, 10) || settings.columns
 
-export const stdtermwidth = columns || termwidth(stdout)
-export const errtermwidth = columns || termwidth(stderr)
+export const stdtermwidth = columns || termwidth(process.stdout)
+export const errtermwidth = columns || termwidth(process.stderr)